# CI: Rails 4+, Ruby 2.4+