--- a/.github/workflows/ci.yaml
+++ b/.github/workflows/ci.yaml
@@ -40,15 +40,15 @@ jobs:
         include:
           - gemfile: Gemfile
             ruby: 2.6
-          - gemfile: gemfiles/rails_3.2.gemfile
-            ruby: 2.3
             bundler: "1"
           - gemfile: gemfiles/rails_4.2.gemfile
             ruby: 2.4
             bundler: "1"
-          # Rails 5.2 is tested with Postgresql below
-          - gemfile: gemfiles/rails_6.1.gemfile
+          - gemfile: gemfiles/rails_5.2.gemfile
             ruby: 2.7
+          # Rails 6.1 is tested with Postgresql below
+          - gemfile: gemfiles/rails_7.0.gemfile
+            ruby: 3.0
           - gemfile: gemfiles/rails_master.gemfile
             ruby: 3.0
           - gemfile: gemfiles/rails_master.gemfile
@@ -100,7 +100,7 @@ jobs:
           --health-timeout 5s
           --health-retries 5
     steps:
-    - run: echo BUNDLE_GEMFILE='gemfiles/rails_5.2_postgresql.gemfile' > $GITHUB_ENV
+    - run: echo BUNDLE_GEMFILE='gemfiles/rails_6.1_postgresql.gemfile' > $GITHUB_ENV
     - run: echo DATABASE='POSTGRESQL' > $GITHUB_ENV
     - run: echo PGPASSWORD='postgres' > $GITHUB_ENV
     - uses: actions/checkout@v2