--- a/Gemfile
+++ b/Gemfile
@@ -8,14 +8,10 @@ gem 'stackprof', platform: :ruby
 gem 'pry'
 gem 'pry-stack_explorer', platform: :ruby
 gem 'graphql-batch'
-if RUBY_VERSION >= "2.4"
-  gem 'pry-byebug'
-end
+gem 'pry-byebug'
 
 # Required for running `jekyll algolia ...` (via `rake site:update_search_index`)
 group :jekyll_plugins do
-  if RUBY_VERSION >= "2.3"
-    gem 'jekyll-algolia', '~> 1.0'
-  end
+  gem 'jekyll-algolia', '~> 1.0'
   gem 'jekyll-redirect-from'
 end