--- a/gemfiles/rails_7.0.gemfile
+++ b/gemfiles/rails_7.0.gemfile
@@ -10,5 +10,6 @@ gem "rails", "~> 7.0.0.alpha2", require: "rails/all"
 gem "sqlite3", "~> 1.4", platform: :ruby
 gem "activerecord-jdbcsqlite3-adapter", platform: :jruby
 gem "sequel"
+gem "evt"
 
 gemspec path: "../"