--- a/graphql.gemspec
+++ b/graphql.gemspec
@@ -13,7 +13,7 @@ Gem::Specification.new do |s|
   s.authors     = ["Robert Mosolgo"]
   s.email       = ["rdmosolgo@gmail.com"]
   s.license     = "MIT"
-  s.required_ruby_version = ">= 2.2.0" # bc `.to_sym` used on user input
+  s.required_ruby_version = ">= 2.4.0" # bc `.to_sym` used on user input
   s.metadata    = {
     "homepage_uri" => "https://graphql-ruby.org",
     "changelog_uri" => "https://github.com/rmosolgo/graphql-ruby/blob/master/CHANGELOG.md",
@@ -34,14 +34,14 @@ Gem::Specification.new do |s|
   s.add_development_dependency "minitest-reporters", "~>1.0"
   s.add_development_dependency "racc", "~> 1.4"
   s.add_development_dependency "rake", "~> 12"
-  s.add_development_dependency "rubocop", "0.68" # for Ruby 2.2 enforcement
+  s.add_development_dependency "rubocop", "1.12" # for Ruby 2.4 enforcement
   # required for upgrader
   s.add_development_dependency "parser"
   # website stuff
   s.add_development_dependency "jekyll"
   s.add_development_dependency "yard"
-  s.add_development_dependency "jekyll-algolia" if RUBY_VERSION >= '2.4.0'
-  s.add_development_dependency "jekyll-redirect-from" if RUBY_VERSION >= '2.4.0'
+  s.add_development_dependency "jekyll-algolia"
+  s.add_development_dependency "jekyll-redirect-from"
   s.add_development_dependency "m", "~> 1.5.0"
   s.add_development_dependency "webrick"
 end