--- a/graphql.gemspec
+++ b/graphql.gemspec
@@ -34,7 +34,7 @@ Gem::Specification.new do |s|
   s.add_development_dependency "minitest-reporters", "~>1.0"
   s.add_development_dependency "racc", "~> 1.4"
   s.add_development_dependency "rake", "~> 12"
-  s.add_development_dependency "rubocop", "~> 1.12" # for Ruby 2.4 enforcement
+  s.add_development_dependency "rubocop", "1.12" # for Ruby 2.4 enforcement
   # required for upgrader
   s.add_development_dependency "parser"
   # website stuff