--- a/lib/graphql/language/lexer.rl
+++ b/lib/graphql/language/lexer.rl
@@ -121,10 +121,6 @@
 module GraphQL
   module Language
     module Lexer
-      if !String.method_defined?(:match?)
-        using GraphQL::StringMatchBackport
-      end
-
       def self.tokenize(query_string)
         run_lexer(query_string)
       end