--- a/lib/graphql/name_validator.rb
+++ b/lib/graphql/name_validator.rb
@@ -1,10 +1,6 @@
 # frozen_string_literal: true
 module GraphQL
   class NameValidator
-    if !String.method_defined?(:match?)
-      using GraphQL::StringMatchBackport
-    end
-
     VALID_NAME_REGEX = /^[_a-zA-Z][_a-zA-Z0-9]*$/
 
     def self.validate!(name)