--- a/lib/graphql/schema/find_inherited_value.rb
+++ b/lib/graphql/schema/find_inherited_value.rb
@@ -1,3 +1,4 @@
+# frozen_string_literal: true
 module GraphQL
   class Schema
     module FindInheritedValue