--- a/lib/graphql/schema/validator/numericality_validator.rb
+++ b/lib/graphql/schema/validator/numericality_validator.rb
@@ -1,3 +1,4 @@
+# frozen_string_literal: true
 module GraphQL
   class Schema
     class Validator