As per https://github.com/rmosolgo/graphql-ruby/issues/3596, I opened a PR with the CI changes needed to support at a minimum
* Ruby 2.4+
* Rails 4+

And prepare the way to support the upcoming Rails 7.

I took the opportunity to make all the changes needed in all the parts of the CI system I could find, like removing `RUBY_VERSION >= '2.4.0'`conditionals, updating rubocop version, and removing some old string refinements that shouldn't be needed anymore.

Removing support from these very old Rails and Ruby versions should speed up development moving forward.

Rails 4 and Ruby 2.4 themselves are still somewhat old and both EOL, but support for those ones is kept for a while longer

This PR can stay open for a while, for people to discuss whether they still need those versions or not.

Personally, I find it hard to believe that there are still people out there who are still using the latest `graphql-ruby` version and always keeping it updated while still using Rails 3 or Ruby 2.2. I'd argue that changing from `define` based to Class based API back in `1.8`, even with the migrator script, took a lot of effort, and anyone who can invest that effort can also invest the effort to keep their Ruby and Rails versions updated. Maybe I'm wrong on this, so let's see what other people have to say.

Also, it's not like GraphQL would stop working for them. Assuming this PR makes it to a `1.13` release, they'd still be able to use everything up to `1.12` and use later versions at their own risk, just like they are using unsupported Ruby and Rails versions that don't receive security fixes at their own risk, anyway.

Tests all worked on my local machine using the ruby versions specified in the `ci.yaml`