@rmosolgo 2021-10-20
>  Wouldn't it make more sense to only support ruby versions that aren't EOL?

In a perfect world, everyone would be running recent Ruby versions. But in practice, some people run old Ruby versions. I'd rather not boot anyone from GraphQL-Ruby if I can help it.

I'm open to dropping support for some old Ruby versions in 1.13.0, but I haven't put a lot of thought into it yet.