@rmosolgo 2021-11-18
Well, shoot. I wanted to remove the string refinement used in the lexer, but in order to rebuild it properly, I had to update Ragel. I added a benchmark task to make sure there wasn't a regression... but there _was_: 

```
# Before, Ragel 7.0.0.9
Warming up --------------------------------------
scan - introspection   110.000  i/100ms
parse - introspection
                        82.000  i/100ms
    scan - fragments   191.000  i/100ms
   parse - fragments   132.000  i/100ms
    scan - big query    18.000  i/100ms
   parse - big query    13.000  i/100ms
Calculating -------------------------------------
scan - introspection      1.087k (± 3.5%) i/s -      5.500k in   5.068137s
parse - introspection
                        809.097  (± 2.1%) i/s -      4.100k in   5.069787s
    scan - fragments      1.882k (± 2.2%) i/s -      9.550k in   5.076049s
   parse - fragments      1.357k (± 1.5%) i/s -      6.864k in   5.061174s
    scan - big query    186.115  (± 2.1%) i/s -    936.000  in   5.031735s
   parse - big query    136.186  (± 3.7%) i/s -    689.000  in   5.066171s

   
# After, Ragel 7.0.4
Warming up --------------------------------------
scan - introspection    66.000  i/100ms
parse - introspection
                        52.000  i/100ms
    scan - fragments   109.000  i/100ms
   parse - fragments    87.000  i/100ms
    scan - big query    10.000  i/100ms
   parse - big query     7.000  i/100ms
Calculating -------------------------------------
scan - introspection    647.735  (± 2.0%) i/s -      3.300k in   5.096960s
parse - introspection
                        532.360  (± 2.3%) i/s -      2.704k in   5.081807s
    scan - fragments      1.048k (± 2.6%) i/s -      5.341k in   5.101512s
   parse - fragments    854.413  (± 2.2%) i/s -      4.350k in   5.093800s
    scan - big query    103.480  (± 1.9%) i/s -    520.000  in   5.026922s
   parse - big query     86.099  (± 3.5%) i/s -    434.000  in   5.049648s
```

So, I guess I _don't_ want to update Ragel. I'll roll back that change and edit the generated file by hand 😖